Add files via upload
This is an algorithm that receives (or not) data about animals admitted to a veterinary clinic. It stores the youngest and oldest ages, and calculates the percentage of a certain type of animal admitted. If no patients were admitted that day, it displays a message about it. Alternative and repetitive structures were incorporated. Algorithm created for a university practical exercise. This algorithm has two ways (and two files) of being executed, both of them very similar but with some differences in structure.